+++ source/module_io/json_output/init_info.cpp
@@ -32,7 +32,7 @@ void gen_init(UnitCell* ucell)
     {
         std::string label = ucell->atoms[it].label;
         int atom_number = ucell->atoms[it].na;
-        int number = ucell->atoms[it].ncpp.zv;
+        double number = ucell->atoms[it].ncpp.zv;
 
         nelec_total += ucell->atoms[it].ncpp.zv * ucell->atoms[it].na;
         AbacusJson::add_json({"init", "natom_each_type", label}, atom_number, false);
